me03 #29942 Address reviewer findings: private const, EMPTY, signature, productIds IN-filter, label recovery comment, @Builder, overpick guard, intValueExact
1. Make MSG_MASS_PRINTING_NOT_ENABLED private (was package-visible, no test uses it).
2. Add MassPrintingResult.EMPTY constant; use it in scan().
3. Simplify retainSelfPackedProducts(ImmutableSet, Map) → retainSelfPackedProducts(Collection<Product>).
4. PackageableQuery: add productIds (ImmutableSet, Builder.Default empty) + push IN-filter into PackagingDAO,
   removing Java-side filter in loadPackageablesGroupedByProduct. Blast radius: 0 callers set productId on
   PackageableQuery builder outside this PR; PackagingDAO precedence: productIds takes priority over productId.
5. Label-print exception: add comment explaining recovery mechanism (labelPrintFailures in ProductResult →
   JSON response → operator reprints manually); keeping best-effort catch is intentional.
6. ScheduleSelection: convert to @Value + @Builder via static factory ScheduleSelection.of(...).
7. Pick events (CU + TU paths): set checkIfAlreadyPacked(true) explicitly + comments documenting
   that cappedQtyInt is already demand-capped by FIFO selectSchedulesFifo, preventing overpick at source.
8. IHUProductStorage.getQtyAsInt(): use getQty().intValueExact() instead of getQty().toBigDecimal().intValueExact().

https://github.com/metasfresh/me03/issues/29942

+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/storage/IHUProductStorage.java
@@ -25,6 +25,6 @@ public interface IHUProductStorage extends IProductStorage, IHUAware
 	 */
 	default int getQtyAsInt()
 	{
-		return getQty().toBigDecimal().intValueExact();
+		return getQty().intValueExact();
 	}
 }
